+++ R/kNNdist.R
@@ -30,7 +30,8 @@
 #'
 #' @param x the data set as a matrix of points (Euclidean distance is used) or
 #' a precalculated [dist] object.
-#' @param k number of nearest neighbors used for the distance calculation.
+#' @param k number of nearest neighbors used for the distance calculation. For
+#' `kNNdistplot()` also a range of values for `k` or `minPts` can be specified.
 #' @param minPts to use a k-NN plot to determine a suitable `eps` value for [dbscan()],
 #'    `minPts` used in dbscan can be specified and will set `k = minPts - 1`.
 #' @param all should a matrix with the distances to all k nearest neighbors be
@@ -61,6 +62,10 @@
 #' ## The knee is visible around a distance of .7
 #' kNNdistplot(iris, k = 4)
 #'
+#' ## Look at all k-NN distance plots for a k of 1 to 10
+#' ## Note that k-NN distances are increasing in k
+#' kNNdistplot(iris, k = 1:10)
+#'
 #' cl <- dbscan(iris, eps = .7, minPts = 5)
 #' pairs(iris, col = cl$cluster + 1L)
 #' ## Note: black points are noise points
@@ -81,11 +86,20 @@ kNNdistplot <- function(x, k, minPts, ...) {
   if (missing(k))
     k <- minPts - 1
 
+  if (length(k) == 1) {
   kNNdist <- sort(kNNdist(x, k, ...))
   plot(
     sort(kNNdist),
     type = "l",
     ylab = paste(k, "-NN distance", sep = ""),
-    xlab = "Points (sample) sorted by distance"
+    xlab = "Points sorted by distance"
   )
+
+  } else {
+    knnds <- sapply(k, FUN = function(i) sort(kNNdist(x, i, ...)))
+
+    matplot(knnds, type = "l", lty = 1,
+            ylab = paste("k-NN distance", sep = ""),
+            xlab = "Points sorted by distance")
+  }
 }